Refactored the functions
